Furthur Refactoring
